+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [IN PROGRESS] C MEX API:
 
+  * mxSetField, mxSetFieldByNumber, mxGetFieldNumber functions.
+
   * mexGetVariable, mexGetVariablePtr, mexPutVariable functions.
 
   * mexMakeArrayPersistent, mexMakeMemoryPersistent functions.